arch/arm: Fix building for CPUs without any FPU
When support for hardware FPU (aka hardfp) was added, the float abi
selected for the soft case was wrong. The ABI should have been 'soft'
instead of 'softfp'. The 'softfp' ABI is compatible with the 'soft'
ABI, however it use the FPU instead of the library calls. This does't
work on CPU without any FPU.

To fix this mess we rename the Kconfig symbols to allow all 3 modes to
be selected, and we mark both softfp and hardfp mode as using an FPU.

Change-Id: I5b1579500e71c32e29e81860eb46355826adc5fa
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1124
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>