net/lldpd: Fix building the SNMP support
The patch to fix the net-snmp detection was incomplete, as a result
the SNMP support code wasn't compiled in. This was because USE_SNMP
wasn't set after the library had been detected.

We also need to manually add libnetsnmpmibs to the LIBS because it
isn't part of the libraries returned by pkg-config.

Change-Id: Ic637ad4bee9d2c88db85c610f60b622284d10bc8
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1468
Reviewed-by: Julian Scheel <julian@jusst.de>