build/rootfs: Allow building the debugfs for the debug symbols
Rework Makefile.rootfs to allow creating another squashfs image that
contains the debug symbols. The resulting image can then be mounted in
/debug for use with gdb.

As the debugfs can potentially be very large, it is not built per
default. To build it the user need to set BUILD_DEBUGFS=y. To skip
building the rootfs one can set BUILD_ROOTFS=n.

Change-Id: I59ee6c4c8d315dbe721aefb3e0cd01e744690431
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1532
Reviewed-by: Julian Scheel <julian@jusst.de>