build: Always reconfigure autoconf packages built from sources
Source link are most commonly used with git checkouts. Such a checkout
need a reconfigure, even if the normal package doesn't need it.

This patch enable secondary expension in the makefile as $(local-src)
isn't defined when autotools.mk is parsed. This shouldn't cause any
problem, but to make sure I first tested this in my local tree for a
few months before submitting it.

Change-Id: I88b80afeac888104c8cbb31f92a7228758590b08
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1747
Reviewed-by: Hendrik Lippek <hendrik.lippek@avionic-design.de>