build-tools: Fix a few missing pruning of the .la
When building libraries with libtool we must always make sure to
remove all the .la. If the .la are left around libtool then fail to
link some target programs as it will prefer the libraries from the
build tools over the one in the sysroot, although they are for the
wrong target.

Change-Id: Ieed92d9d3cc66321f8536b454ef20a0a01a5346f
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1748
Reviewed-by: Hendrik Lippek <hendrik.lippek@avionic-design.de>