packages/gtk3: Fix build on system without automake 14
The patch to fix the compilation of the icon cache tool changes a
Makefile.am without updating the relevant Makefile.in. This triggers
autoreconf, however it requires automake 14 and we only have automake 13.
As updating automake to version 14 breaks many packages we take the
simple route and add a patch to also patch Makefile.in to avoid the
need to reconfigure.

Change-Id: I502afa06c86fb4e329ccf198da971dbadedf8449
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1768
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>