packages/synergy: Fix reconnection after resolv.conf changed
The synergy client fails to connect to its server when it is
started while resolv.conf isn't properly configured. This might happen
when synergy is started before the system obtained its DHCP lease.

Add a patch that call res_init() when the name resolution failed to
reload resolv.conf and solve this issue.

Change-Id: Ied5053dc0b93878f18429165c6ce9164b914692d
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1772
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>