build: deps: Fix the dependency list generation
The dependency list generation goes thru the list of reverse depencies
of each package. It then adds this package as dependency to all
packages that need it.

However it currently doesn't check if any of the reverse dependencies is
actually needed. So if a package A can use package B when condition C is
met, package B is always added to the dependencies of A even if the
condition C isn't met.

This first creates some dependency that shouldn't exists, but more
importantly it prevents virtual packages from working properly. Virtual
packages should only have a single dependency on the package they
virtualize, however because of this bug they might have several
dependencies.

To fix this problem we re-implement the dependency with proper back
tracking to only add the dependencies that are actually needed.

Change-Id: Ifc9a10e323bdff4efa1ec1c979139fd7179dec56
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1785
Reviewed-by: Julian Scheel <julian@jusst.de>