core/util-linux: Fix PAM switch behaviour
The packages Kconfig contains a switch to enable or disable PAM and hence
login support explicitly. While this is used to conditionally depend on
LIBS_PAM in the Kconfig it is not used in the Makefile to explicitly disable
login with its PAM dependency for the configure step. Instead the Makefile
checked for LIBS_PAM being enabled as a package which might be due to some
other package depending on it. In this case --disable-login was not used
although the config option to enable login support for util-linux was not set.

Besides changing the behaviour from the expected one (not including login
support is expected, but in this constellation it is included) it can break
build depending on the build dependencies of other packages.

Change-Id: I05a0b6ab05718e3b6a9c6d67a760bd70cdd31d5b
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/1889
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>