packages/remote-control: disable -werror for now
Temporary disable -werror for remote-control. After the -werror switch
in remote-control has been fixed, the build is broken since warnings
occur on a regular base. While they were ihnored previously, they now
pose a problem and need to be fiexd. Until that happens, -werror needs
to be disabled to be able to build a platform with remote-control.

Change-Id: Ia34f68de6033faa69b7e3b361a90c3d554c98190
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1920
Reviewed-by: Hendrik Lippek <hendrik.lippek@avionic-design.de>