packages/autotools: Add a common fix for AC_FUNC_MALLOC/MEMCMP
The autoconf checks AC_FUNC_MALLOC and AC_FUNC_MEMCMP are not working
when cross compilling. Up to now packages tackled this issue only when
noticed, many using a slightly different method, some beeing plain
wrong.

To fix this once and for all add the proper conf-vars to
autotools-base.mk.

Change-Id: Ic149edd559f48bb0eecffa91000a8f48ccad3d6f
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1982
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>