packages/dfu-programmer: Add a patch to fix rpl_malloc()
It turned out that the dfu-programmer hang reported in FS#3005
was due to a bug the rpl_malloc() implementation. The fix implemented
in commit 2608940cfcd8 just work-arounded the problem by always
setting ac_cv_func_realloc_0_nonnull=yes. However this is only true
for glibc.

Add a patch that fix the infinite loop in rpl_malloc().

Change-Id: I5be8b2bffe8cbacfb45284a88b171dfce118b296
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1984
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>