build: Allow packages to use version specific patches
If a directory named patches-$(VERSION) exists, it gets used instead of
the default patches directory. This is a first step to allow supporting
several versions of a package. Note that a few packages already
implement a similar mechanism when several versions are needed.See
the xorg-server packages for an example. As packages can still
override $(patchdir) manually they will continue to works as is.

Change-Id: I0aa714552192431e22faa0021a923987c59329dc
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/2145
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>