packages/lldpd: Update to version 0.7.11
Update all the patches to the new version, add libevent to the
dependencies and disable readline. The latter is to get the same build
when the readline package is built or not.

Add the install files for the new lldp control library and its
developments files.

Systemd support has been implemented so we also drop our systemd patch.
However the systemd support has a few bugs for which debian already
has some patches, so we also add those.

Also add /etc to the install for the systemd init files. Finally note that
platforms that used older versions of this package with systemd will
need to drop their lldpd.service and lldpd.socket files.

Change-Id: Ia7d8b679f74ea7bb689c8083dc058f88a72edce0
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/2422
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>