scripts/watch: Fix the href matching to cope with more cases
When looking for an href, only match the whole property when
matching an absolute path. If we only match a relative path, allow the
href to have some extra path components before the match.

This is more an heuristic than anything else but it is working pretty
well for now.

Change-Id: I3015a2522ff64a1eae5862a13471a7810a1b1f79
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/2234
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>