build: Add a rule to build scripts/kconfig/deps
scripts/kconfig/deps is needed to generate depends.mk, however there
is no explicit rule to build it from the top makefile. This is
normally not a problem as it is built as part of kconfig, but in case
it is erased it is nice to have rule that allow rebuilding it.

Change-Id: I906a290291ff8cd4d5b94c5fa6a1193f89df5e8f
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/2236
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>