packages: xorg-server: getevents: Add patch to check inverted tranformation
If a touch position change was reported with only one axis, the X-server tried
to calculate the missing axis via an inverted transformation. But if the base
transformation is not used/valid, this will generate invalid results.
The symptom is that the mouse pointer jumps to the zero position on the
missing axis. Special the Sharp touch (04dd:9570) seems to force this situation.
We now prevent this by checking the return code of the invert function.

Change-Id: Idab9b6816f5b0e76fc1a91f638b5ac10b51b06f5
Signed-off-by: Dirk Leber <dirk.leber@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/3493
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>