build: add a time target for printing build times
Add a per-package 'time' target for printing the build time per package.
When invoked on a platform, all selected packages and their build times
are printed as a list. Time is measured by evaluating .directory and
.install stamps modification dates, the only stamps all packages have in
any case. Also works when rebuilding and cleaning packages before
building them, but not if build steps are invoked independently.

Change-Id: I2662087d1a2bcbe2cd0cff189f30a3a180a486fa
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/4514
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>