packages/web/webkit-gtk: repatriate stray patch
This patch ended up lost and lone since the great merge. Put it back
into the patches directory for 1.10.2 to get it applied again and
prevent stupid systems from crashing the build due to a missing patch
anounced by the series file. Loose the number in the patch file name as
it is pointless in this case.

Change-Id: I1cc26198548f907672bad870f25b42eccdaa56be
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5423
Reviewed-by: Hendrik Lippek <hendrik.lippek@avionic-design.de>