scripts/watch: hint if packages look pinned
Watch files containing numeric regex groups almost always are pinned to
a specific version or exclude a range of newer versions. While there
might be valid reasons at the time, pinning is not always desirable and
prevents displaying possible updates in the future. Make the watch
script display a hint if packages seem to be pinned.

Change-Id: Id498a0efa4f49f5a967127232317cd4cf51707f4
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5512
Reviewed-by: Hendrik Lippek <hendrik.lippek@avionic-design.de>