packages/mm/vlc: add patch for old dvb api compat
The l4t-r16-3 kernel has a DVB API which is older than the current VLC 3
nightly expects in one of its source files. Patch that file to be able
to build. Note that the current upstream nightly has this part already
changed sufficiently to avoid this problem in the future.

Change-Id: I8392fbfd6fad96516ee9a30afa5dfd8ec71277e7
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5607
Reviewed-by: Hendrik Lippek <hendrik.lippek@avionic-design.de>