packages/libs/pcsc/lite: fix broken updates
Updates for this package were done by incrementing the version number,
but the download location requires a file ID to be incremented as well.
Failing to do this resulted in downloading a seemingly updated tarball
with a source tree from the original (old) version. This triggered a
bugfix referring to that tree instead of checking whether the package
really was updated. Another strong hint were identical checksums for all
versions. Fix this mess by updating the checksum for 1.8.16, removing
the one for 1.8.13 and removing the patches for both as they are unneeded.

Change-Id: I78faf8e02631057f5ddf4ffc72b0093579cc779f
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5639
Reviewed-by: Julian Scheel <julian@jusst.de>