scripts/watch: Add a workaround to better cope with directory listings
At some sites, like mozilla.org, directory listings are not working
(ie return a 404) unless the URL has a trailing /. This adds a trivial
heuristic that appends a trailing / to the URL unless the last path
commponent seems to be a file or there is a query part. The heuristic
to detect a file just looks for a . (for file extensions) or a :
(for wiki pages).

This is far from bullet proof but seems to work well with all the
packages used for the medcom-wide platform.

Change-Id: I4f66b19ecf33518375f7f4157c878f6d902b12bf
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6338
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>