packages/build-tools/ca-certificates: add new package
Copy the ca-certicates to the build tools as Python from there needs it
for verifying SSL certificates. This now is a requirement since the
watch script relies on the build tools Python and Python in turn always
tries to verify certificates, from v2.7.9 on. Alternatives would be
changing any https:// watch file URIs or forcing the watch script to
ignore SSL certificates, both of which feels a bit anachronistic in
2016.

Change-Id: I4f0cc67422771cd417e42b614055aa3925a4c948
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6956
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>