scripts/watch: support alternative version separator
Sometimes tarballs have their version separated by an underscore or a
dash instead of a dot, as semantic versions should have. When the
Makefile's VERSION variable does not use the same separator, false
positives for available updates are reported. Support a
'version_separator' parameter in watch files for better version
comparison. There is no change for watch files that do not use it.

Change-Id: I34fc356bd0ae74f57f260bb3c50ef5b488db2305
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7082
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>