packages/autotools-base: use {prefix} for libdir
Do not use a hard-coded path but rather the ${prefix} variable for
libdir. The original intention was to force things to be installed in
lib/ instead of lib64/, but using the ${prefix} variable instead of its
value is still required if the path needs to be patched afterwards. This
is the case with library config-scripts and now works by just altering
the prefix. Change man and info paths as well. Side effects should be
nil.

Change-Id: I8cf3bc808e0c53c1e708d737decd964bb361bc55
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7100
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>