build: Silence useless messages when computing the sha256 checksum
First remove the "check sha256 for download/..." message, it is
useless as the build process already print out that it is doing the
sha256 using the standard build step messages.

Furthermore let sha256sum only report via status code, that avoid both
success and failure messages. The success message is useless in our
case and the failure message is too terse, so we replace it with a
message of our own.

Finally also send stderr to /dev/null to avoid the "no properly
formatted SHA256 checksum lines found" message when no checksum exists
yet for the current version, as well as the extra warning when the
checksums check fails.

Change-Id: I74727095745f9b523c4e3488b00b4ffc35d0f2ed
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7098
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>