packages/build-tools/docbook-xml: relink old versions locally
Make docbook versions from before the current one work by pointing them
to the version that is installed. This requires generating
/etc/xml/docbook as using $(db_catalog) as replacement does not work.
The patch fixes packages using old docbook versions and generators with,
either explicitly or implicitly (xmlto), the --nonet switch. Most don't
make a hassle, but some packages even fail the whole build because of
that.

Change-Id: I9f35c86b5af1d8529b5b35f20d1cd27fbd18fae5
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7116
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>