build: use separate bin dir for target config-scripts
Custom library config-scripts as alternative to pkg-config are installed
into the build-tools' bin/ subdirectory with their paths changed to
sysroot. This works well for target libs but leads to trouble for
libraries which are built for both the build-tools and the target. Since
the respective <lib>-config script gets overwritten and there is no
guarantee packages get built in a specific order, building is now a bit
non-deterministic. Solve this by putting config-scripts meant for the
target into build-tools' bin/target/ subdirectory and prepending the
PATH variable accordingly for non build-tools packages. The only way
this can go wrong is a missing target but an existing build-tools
<lib>-config script, but that's a mere dependency problem.

Change-Id: Ife08c51ffb373aec8905fb57c5cdd232ea6c959f
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7223
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>