packages/build-tools/u-boot: rebuild config options
For build-tools only the u-boot tools are needed, so reduce the
configuration to the absolute minimum. Additionally make FIT and FIT
signature support optional to avoid hard-coded dependencies on dtc and
openssl. The latter is enabled per default now, triggering this whole
change. Enable FIT support as default to avoid breaking defconfigs.

Change-Id: If4a90922b123c4c23079a829dc2d998b262b4c38
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7412
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>