packages/web/webkit-gtk: add option for webkit1 support
The legacy webkit1 API has been dropped from current webkit-gtk
releases. Since the webkit2 API is rather incompatible and porting is
not always an option, add a Kconfig switch to enable webkit1 API support
explicitly. This reverts to version to a release from the 2.4.x branch,
the last to support the legacy API.

Change-Id: I6de1c9f1d811e6c07d52b277657514982ec616c7
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7416
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>