scripts/watch: fix scans failing in subdirs with RCs only
When invoking the watch script scanning for the latest release and
release are collected in version subdirectories, the scan failed when
only RCs or betas but no release were available. This is now fixed by
keeping the full path list, reversing the order and use the next entry
if the current path fails. As a side effect a version ordering bug at
patch level is also fixed. No missed update reports or extra scan
failures compared to the previous watch script version have been
observed.

Change-Id: Icacecf08c697cdd0bb17d05b082bbd4f18da3dd4
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7426
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>