packages/core: fix usbutils vs udev dependencies
The usbutils package depends on udev, not the other way around. Both
systemd and udev supply libudev. Remove the disable_extras config switch
from udev since it is not used anymore.

Note that udev has been merged into systemd and the original package is
discontinued. Gentoo forked udev into eudev to avoid a dependency on
systemd when only udev is needed. The udev package should be replaced
with eudev in the future.

Change-Id: I7de055346604cb71e8f487d0b12d37da42da1c10
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7566
Reviewed-by: Meike Vocke <meike.vocke@avionic-design.de>