packages/build-tools/grep: add new package
Add grep to build-tools to make PBS a bit more independent from the
host. Some hosts appear to have a buggy grep installed, the most recent
example was grep 2.24 failing on a UTF-8 file in shared-mime-info,
breaking the build. Having grep in build-tools allows controlling which
grep exactly is used.

Change-Id: I1d960c9d9e8fd1d8a71563d4a320b965643504ad
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7650
Reviewed-by: Meike Vocke <meike.vocke@avionic-design.de>