packages/core/systemd: preset paths for target binaries
The systemd package cleverly detects paths for binaries used at runtime
on the host and tries to invoke those on the target system. This works
as long as the locations on host and sysroot match, however for example
the util-linux package now puts mount into /bin instead of /usr/bin and
thereby fools systemd's detection. Clean up this mess by presetting
paths.

Change-Id: Ic2f08f3d37d47221ca5306aa761d4a2d2e656b5f
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/8306
Reviewed-by: Meike Vocke <meike.vocke@avionic-design.de>